refactor(dom): extract shared withCDPConnection, remove dead --port (#216)
Three command files (dom/index.ts, form.ts, formInteraction.ts) each
had a near-identical withCDPConnection helper opening direct CDP
WebSockets from the CLI process. Extract into a single module so the
daemon-boundary leak is contained in one place, to be migrated to IPC
as a follow-up.

Also removes the dead '-p, --port' flag from 'bdg dom eval' — the
handler explicitly ignored the CLI value and used session metadata.
Drops PortOptions type (no other consumers).

Net: -127 lines, no behavior change for valid invocations.

Remaining dual-transport work (routing DOM operations through IPC
callCDP rather than a second CDP connection) tracked as part of #216.
